random policy code restructure
