+++ requirements.txt
@@ -1,24 +1,38 @@
+absl-py==0.11.0
 appnope==0.1.2
 argon2-cffi==20.1.0
+astunparse==1.6.3
 async-generator==1.10
+atari-py==0.2.6
 attrs==20.3.0
 backcall==0.2.0
 bleach==3.2.3
+cachetools==4.2.1
+certifi==2020.12.5
 cffi==1.14.4
+chardet==4.0.0
 cloudpickle==1.6.0
 cycler==0.10.0
 decorator==4.4.2
 defusedxml==0.6.0
 entrypoints==0.3
+flatbuffers==1.12
 future==0.18.2
-grpcio==1.35.0
+gast==0.3.3
+google-auth==1.24.0
+google-auth-oauthlib==0.4.2
+google-pasta==0.2.0
+grpcio==1.32.0
 gym==0.18.0
+h5py==2.10.0
+idna==2.10
 ipykernel==5.4.3
 ipython==7.19.0
 ipython-genutils==0.2.0
 ipywidgets==7.6.3
 jedi==0.18.0
 Jinja2==2.11.2
+joblib==1.0.0
 jsonschema==3.2.0
 jupyter==1.0.0
 jupyter-client==6.1.11
@@ -26,19 +40,26 @@ jupyter-console==6.2.0
 jupyter-core==4.7.0
 jupyterlab-pygments==0.1.2
 jupyterlab-widgets==1.0.0
+Keras-Preprocessing==1.1.2
 kiwisolver==1.3.1
+Markdown==3.3.3
 MarkupSafe==1.1.1
 matplotlib==3.3.4
 mistune==0.8.4
 mlagents-envs==0.18.1
+mpi4py==3.0.3
 nbclient==0.5.1
 nbconvert==6.0.7
 nbformat==5.1.2
 nest-asyncio==1.5.1
 notebook==6.2.0
 numpy==1.18.5
+oauthlib==3.1.0
 -e git+https://github.com/Unity-Technologies/obstacle-tower-env.git@d2450cd5fea92fff0a881e858a9a44f53a440d19#egg=obstacle_tower_env
+opencv-python==4.5.1.48
+opt-einsum==3.3.0
 packaging==20.9
+pandas==1.2.1
 pandocfilters==1.4.3
 parso==0.8.1
 pexpect==4.8.0
@@ -48,23 +69,39 @@ prometheus-client==0.9.0
 prompt-toolkit==3.0.14
 protobuf==3.14.0
 ptyprocess==0.7.0
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
 pycparser==2.20
 pyglet==1.5.0
 Pygments==2.7.4
 pyparsing==2.4.7
 pyrsistent==0.17.3
 python-dateutil==2.8.1
+pytz==2020.5
 PyYAML==5.4.1
 pyzmq==22.0.2
 qtconsole==5.0.2
 QtPy==1.9.0
-scipy==1.6.0
+requests==2.25.1
+requests-oauthlib==1.3.0
+rsa==4.7
+scipy==1.4.1
 Send2Trash==1.5.0
 six==1.15.0
+tensorboard==2.2.2
+tensorboard-plugin-wit==1.8.0
+tensorflow==2.2.0
+tensorflow-estimator==2.2.0
+termcolor==1.1.0
 terminado==0.9.2
 testpath==0.4.4
+torch==1.7.1
 tornado==6.1
 traitlets==5.0.5
+typing-extensions==3.7.4.3
+urllib3==1.26.3
 wcwidth==0.2.5
 webencodings==0.5.1
+Werkzeug==1.0.1
 widgetsnbextension==3.5.1
+wrapt==1.12.1
